Move EvalSnapshot into its own module
